rename dir to build-and-import
