+++ packages/prerender/src/build-and-import/rollupPlugins/__tests__/__fixtures__/cell/cell-with-default-export/code.js
