+++ packages/prerender/src/build-and-import/rollupPlugins/__tests__/__fixtures__/route-auto-loader/failure/web/src/pages/HomePage/useHomePage.tsx
