+++ packages/prerender/src/build-and-import/rollupPlugins/__tests__/__fixtures__/viteDistDir/client-build-manifest.json
