+++ packages/prerender/src/build-and-import/rollupPlugins/__tests__/__fixtures__/viteSrcDir/components/Post/Posts/image1.jpg
