+++ packages/prerender/src/build-and-import/rollupPlugins/__tests__/__fixtures__/viteSrcDir/pages/HomePage/small.jpg
