+++ packages/prerender/src/build-and-import/rollupPlugins/__tests__/__fixtures__/viteSrcDir/pdf/invoice.pdf
