+++ packages/prerender/src/build-and-import/rollupPlugins/rollup-plugin-cedarjs-external.ts
