+++ packages/prerender/src/build-and-import/rollupPlugins/rollup-plugin-cedarjs-inject-file-globals.ts
